requirements.txt
watchdog>=3.0.0
boto3>=1.28.0
Pillow>=10.0.0
PyYAML>=6.0
python-dotenv>=1.0.0
